Change variable names. Set query param to detect language from to 'lang'. Move i18n.ts to 'services' folder
